requirements.txt
alembic==1.9.1
amqp==5.1.1
appdirs==1.4.4
async-timeout==4.0.2
beautifulsoup4==4.11.1
billiard==3.6.4.0
celery==5.2.7
celery-redbeat==2.0.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==39.0.0
Flask==2.2.2
Flask-Migrate==4.0.1
Flask-SQLAlchemy==3.0.2
frozendict==2.3.4
greenlet==2.0.1
html5lib==1.1
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
lxml==4.9.2
Mako==1.2.4
MarkupSafe==2.1.1
multitasking==0.0.11
mysql-connector-python==8.0.31
numpy==1.24.1
pandas==1.5.2
prompt-toolkit==3.0.36
protobuf==3.20.1
pycparser==2.21
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7
redis==4.4.1
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.46
tenacity==8.1.0
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.2
yfinance==0.2.4
zipp==3.11.0
